# Fix an issue in `cupyx.scipy.sparse.linalg.eigsh` with CUDA 9.2