This PR fixes an issue in `cupyx.scipy.sparse.linalg.eigsh` with CUDA 9.2.
https://github.com/cupy/cupy/pull/4229#issuecomment-722241735