@toslunar 2020-11-06
Speed on V100.  I used https://github.com/cupy/cupy/pull/4138#issuecomment-717158904.

size | nnz | nnz per row | k | scipy.eigsh (sec) | cupyx.eigsh before (sec) | cupyx.eigsh after (sec)
-----|-----|-------------|---|-------------------|------------------|--
5000 | 248717 | 49.7 | 3 | 0.081 | 0.061 | 0.058
5000 | 248717 | 49.7 | 6 | 0.120 | 0.135 | 0.123
5000 | 248717 | 49.7 | 12 | 0.184 | 0.312 | 0.279
5000 | 478250 | 95.7 | 3 | 0.121 | 0.099 | 0.093
5000 | 478250 | 95.7 | 6 | 0.163 | 0.130 | 0.119
5000 | 478250 | 95.7 | 12 | 0.269 | 0.332 | 0.308
5000 | 967850 | 193.6 | 3 | 0.278 | 0.105 | 0.105
5000 | 967850 | 193.6 | 6 | 0.354 | 0.149 | 0.137
5000 | 967850 | 193.6 | 12 | 0.519 | 0.323 | 0.292
10000 | 485705 | 48.6 | 3 | 0.141 | 0.067 | 0.066
10000 | 485705 | 48.6 | 6 | 0.178 | 0.111 | 0.104
10000 | 485705 | 48.6 | 12 | 0.420 | 0.325 | 0.310
10000 | 980932 | 98.1 | 3 | 0.344 | 0.113 | 0.109
10000 | 980932 | 98.1 | 6 | 0.406 | 0.154 | 0.134
10000 | 980932 | 98.1 | 12 | 0.704 | 0.381 | 0.333
10000 | 1972756 | 197.3 | 3 | 0.765 | 0.142 | 0.140
10000 | 1972756 | 197.3 | 6 | 0.842 | 0.162 | 0.150
10000 | 1972756 | 197.3 | 12 | 1.592 | 0.458 | 0.416
20000 | 975180 | 48.8 | 3 | 0.660 | 0.132 | 0.128
20000 | 975180 | 48.8 | 6 | 0.873 | 0.213 | 0.193
20000 | 975180 | 48.8 | 12 | 1.086 | 0.357 | 0.333
20000 | 1925696 | 96.3 | 3 | 0.879 | 0.232 | 0.211
20000 | 1925696 | 96.3 | 6 | 1.006 | 0.167 | 0.147
20000 | 1925696 | 96.3 | 12 | 1.643 | 0.383 | 0.363
20000 | 4161452 | 208.1 | 3 | 1.394 | 0.103 | 0.100
20000 | 4161452 | 208.1 | 6 | 2.301 | 0.242 | 0.211
20000 | 4161452 | 208.1 | 12 | 3.144 | 0.415 | 0.389

<details>

```
OS                           : Linux-4.15.0-58-generic-x86_64-with-glibc2.27
CuPy Version                 : 9.0.0a1
NumPy Version                : 1.19.2
SciPy Version                : 1.5.2
Cython Build Version         : 0.29.21
Cython Runtime Version       : 0.29.21
CUDA Root                    : /usr/local/cuda
CUDA Build Version           : 10010
CUDA Driver Version          : 11010
CUDA Runtime Version         : 10010
cuBLAS Version               : 10201
cuFFT Version                : 10101
cuRAND Version               : 10101
cuSOLVER Version             : (10, 2, 0)
cuSPARSE Version             : 10300
NVRTC Version                : (10, 1)
Thrust Version               : 100906
CUB Build Version            : 100800
cuDNN Build Version          : 7500
cuDNN Version                : 7500
NCCL Build Version           : 2703
NCCL Runtime Version         : 2703
cuTENSOR Version             : 10200
Device 0 Name                : Tesla V100-SXM2-16GB
Device 0 Compute Capability  : 70
```

code:
```python
import functools
import math
import numpy
import scipy
import scipy.sparse as sparse
import scipy.sparse.linalg as linalg
import cupy
import cupyx
import cupyx.scipy.sparse as cp_sparse
import cupyx.scipy.sparse.linalg as cp_linalg

dtype = numpy.float64
n_repeat = 10
n_warmup = 1


def make_sp_matrix(m, target):
    max_density = 0.0001
    min_density = 0
    while True:
        density = (min_density + max_density) / 2
        a = sparse.random(m, m, density=density, format='csr', dtype=dtype)
        a = a * a.conj().T
        nnz_per_row = a.nnz / m
        if nnz_per_row < target * 0.5:
            max_density *= math.sqrt(target / nnz_per_row)
        elif nnz_per_row < target * 0.95:
            min_density = density
        elif nnz_per_row > target * 1.05:
            max_density = density
        else:
            return a


def get_time(perf):
    cpu_time = perf.cpu_times.mean()
    gpu_time = perf.gpu_times.mean()
    return max(cpu_time, gpu_time)


def run(sp_a, k):
    times = []
    perf = cupyx.time.repeat(linalg.eigsh,
                             (sp_a, k),
                             n_repeat=n_repeat, n_warmup=n_warmup)
    times.append(get_time(perf))
    for gpu in [True, False]:
        eigsh = functools.partial(cp_linalg.eigsh, gpu)
        eigsh.__name__ = 'eigsh'
        perf = cupyx.time.repeat(eigsh,
                                 (cp_sparse.csr_matrix(sp_a), k),
                                 n_repeat=n_repeat, n_warmup=n_warmup)
        times.append(get_time(perf))
    return times


print('size | nnz | nnz per row | k | scipy.eigsh (sec) | cupyx.eigsh before (sec) | cupyx.eigsh after (sec)')
print('-----|-----|-------------|---|-------------------|------------------|--')
for m in 5000, 10000, 20000:
    for nnz_per_m in 50, 100, 200:
        a = make_sp_matrix(m, nnz_per_m)
        for k in 3, 6, 12:
            times = run(a, k)
            print('{} | {} | {:.1f} | {} | {:.3f} | {:.3f} | {:.3f}'
                  ''.format(m, a.nnz, a.nnz / m, k, times[0], times[1], times[2]))
```

with diff

```diff
diff --git a/cupyx/scipy/sparse/linalg/_eigen.py b/cupyx/scipy/sparse/linalg/_eigen.py
index 5a920b6ce..dc9ea7492 100644
--- a/cupyx/scipy/sparse/linalg/_eigen.py
+++ b/cupyx/scipy/sparse/linalg/_eigen.py
@@ -2,8 +2,9 @@ import numpy
 import cupy


-def eigsh(a, k=6, *, which='LM', ncv=None, maxiter=None, tol=0,
+def eigsh(gpu, a, k=6, *, which='LM', ncv=None, maxiter=None, tol=0,
           return_eigenvectors=True):
+    _eigsh_solve_ritz = _eigsh_solve_ritz_gpu if gpu else _eigsh_solve_ritz_cpu
     """Finds ``k`` eigenvalues and eigenvectors of the real symmetric matrix.

     Solves ``Ax = wx``, the standard eigenvalue problem for ``w`` eigenvalues
@@ -131,7 +132,26 @@ def _eigsh_lanczos_update(A, V, alpha, beta, i_start, i_end):
     return u


-def _eigsh_solve_ritz(alpha, beta, beta_k, k, which):
+def _eigsh_solve_ritz_gpu(alpha, beta, beta_k, k, which):
+    t = cupy.diag(alpha)
+    t = t + cupy.diag(beta[:-1], k=1)
+    t = t + cupy.diag(beta[:-1], k=-1)
+    if beta_k is not None:
+        t[k, :k] = beta_k
+        t[:k, k] = beta_k
+    w, s = cupy.linalg.eigh(t)
+
+    # Pick-up k ritz-values and ritz-vectors
+    if which == 'LA':
+        idx = cupy.argsort(w)
+    elif which == 'LM':
+        idx = cupy.argsort(cupy.absolute(w))
+    wk = w[idx[-k:]]
+    sk = s[:, idx[-k:]]
+    return wk, sk
+
+
+def _eigsh_solve_ritz_cpu(alpha, beta, beta_k, k, which):
     # Note: This is done on the CPU, because there is an issue in
     # cupy.linalg.eigh with CUDA 9.2, which can return NaNs. It will has little
     # impact on performance, since the matrix size processed here is not large.
```